+++ vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
     test: {
-        include: ['extensions/*/src/**/*.test.ts'],
+        include: ['extensions/*/src/**/*.test.ts', 'scripts/*.test.js'],
         environment: 'node'
     }
 });